+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+v5.7.14
+----------
+ * Stop writing is_blocked and is_stopped
+
+v5.7.13
+----------
+ * Read from contact.status intead of is_stopped/is_blocked
+ * Implement saving of zendesk ticket files as attachments
+ * Abstract S3 code so tests and dev envs can use file storage
+
 v5.7.12
 ----------
  * Fix inserting channel logs and add test